requirements.txt
Flask
requests
PyPDF2
PyCryptodome
PyMuPDF
python-dotenv
tqdm
Pillow
google-generativeai 